requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
causal-learn==0.1.4.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.1
contourpy==1.3.3
cryptography==46.0.3
cvxpy==1.7.5
cycler==0.12.1
Cython==3.2.3
distro==1.9.0
dowhy==0.14
et_xmlfile==2.0.0
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
momentchi2==0.1.8
mpmath==1.3.0
narwhals==2.14.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
openai==2.14.0
openpyxl==3.1.5
osqp==1.0.5
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
plotly==6.5.0
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydot==4.0.1
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.15.3
scs==3.2.9
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
statsmodels==0.14.6
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
